medical/serializers.py
from rest_framework import serializers
from .models import Alergia, Enfermedad, HistorialMedico

class AlergiaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Alergia
        fields = '__all__'

class EnfermedadSerializer(serializers.ModelSerializer):
    class Meta:
        model = Enfermedad
        fields = '__all__'